+++ client/chatbox.js
@@ -1,0 +1,13 @@
+'use strict'
+
+import React from 'react'
+
+const Chatbox = () => (
+  <div>
+    <iframe
+    width="500"
+    height="600"
+    src="https://console.api.ai/api-client/demo/embedded/b77433ff-768e-41ff-a8a7-24fe2a5219b8" />
+  </div>
+)
+export default Chatbox
