+++ client/root.js
@@ -1,6 +1,7 @@
 import React from 'react'
 import About from './about'
 import Chat from './chat'
+import Chatbox from './chatbox'
 import Slack from './slackbutton'
 
 const Root = () => (
