+++ public/styles.css
@@ -24,3 +24,8 @@ body {
   font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", Helvetica, Arial, sans-serif, "Apple Color Emoji", "Segoe UI Emoji", "Segoe UI Symbol";
   color: thunder
 }
+
+#textbox {
+  height: 200;
+  width: 200;
+}
